network optimization: ensure objective values are float32

+++ src/miv_simulator/optimize_network.py
@@ -439,13 +439,13 @@ def compute_objectives(local_features, operational_config, opt_targets):
         )
 
         all_features_dict[f"{pop_name} mean fraction active per time bin"] = (
-            mean_fraction_active_per_bin
+            float(mean_fraction_active_per_bin)
         )
         all_features_dict[f"{pop_name} std fraction active per time bin"] = (
-            std_fraction_active_per_bin
+            float(std_fraction_active_per_bin)
         )
-        all_features_dict[f"{pop_name} fraction active"] = fraction_active
-        all_features_dict[f"{pop_name} firing rate"] = mean_rate
+        all_features_dict[f"{pop_name} fraction active"] = float(fraction_active)
+        all_features_dict[f"{pop_name} firing rate"] = float(mean_rate)
 
         rate_constr = mean_rate if mean_rate > 0.0 else -1.0
         constraints.append(rate_constr)
@@ -470,7 +470,7 @@ def compute_objectives(local_features, operational_config, opt_targets):
         features.append(feature_val)
 
     result = (
-        np.asarray(objectives),
+        np.asarray(objectives, dtype=np.float32),
         np.array([tuple(features)], dtype=np.dtype(feature_dtypes)),
         np.asarray(constraints, dtype=np.float32),
     )
